map proto & debug Window
Got a working map prototype and created a detachable debug window.
hinsert has a bug right now. echoes will not work inside triggers.
To get debug info in a window. Run debug on. Than rather echo troubleshooting information to screen, run debugToDisplay("Information")
It will appear on the dock debug window.
Window is created when debug on is run. It's off to the right. I forgot to have it recall position.