tecMap V1 Completed
Exit icons now properly position themselves.
As the window resizes the tecMap and compass stay square.
Later releases we can create an option to not do this. To me things look a bit off when they are not square.
Created full theme support for the exit icons.
png images were only needed for angled exits. They are in the ims.mpackage file.