tecMap V1 Completed (#7)
* tecMap V1 Completed

Exit icons now properly position themselves.
As the window resizes the tecMap and compass stay square.
Later releases we can create an option to not do this. To me things look a bit off when they are not square.
Created full theme support for the exit icons.
png images were only needed for angled exits. They are in the ims.mpackage file.

* Map Bug Fix

In one area of the game with many rooms. It reached past the max exit limit of 100.
Set the new limit to 150.
Renamed the GUI tab to TEC Map
Created a list of things to do in future releases.