Completed Updates
Complete update package.
Downloads a file to check if most up to date client is in use.
If an update is needed the user is prompted at login.
When the user runs the update a file containing a table listing all scripts the project uses downloads. This allows for the addition of more script files in the future.
Using that file all scripts are downloaded. Old packages are uninstalled and new ones are installed.
This script requires a file containing all scripts to be installed and a file containing version names,
Aliases 'tecclient save versions' and 'tecclient save script names' have been created to assist in the creation of these files.
These two files and all scripts must be in a static releases folder on github. It is  possible to allow this URL to change from version to version, though human error would likely result in a version failing to update.