Select Password & Clear (#44)
Selects the line the password is in. Finds the password in that line, and replaces it.
Made certain the command being passed by sysDataSendRequest gets cleared. We expect it to be as it is a local of function controlInput though. Also clearing out command variable, if we do not it will hold our password until another command is entered by the user.
Security update.