Powershell Script To Install Windows (#42)
Powershell script installs mudlet if it is not.
Installed mudlet profile for the eternal city if it is not.
It backups up the settings folder if an existing mudlet profile for The Eternal City exists. With the intenion of restoring user created scripts. After installing the fresh profile.
Starts mudlet after installation completion.
Using PS2EXE to convert powershell script into an exe.
https://gallery.technet.microsoft.com/scriptcenter/PS2EXE-GUI-Convert-e7cb69d5