RoomCharDetectionUpdate
Scripts
ln 234, 235, 4185, 4201 changed variables to have color at the end of their name.
Triggers
ln 645, 652, 678, 679,  changed characterEngaged to characterStatus. Did this because we are keeping track of statuses now.
ln 686 fully checks if character is engaged
ln 689 - 696 if character is unconscious is prefices their name with `KO: `
ln 698 699 updated for change in color variable
lns 719 - 725 Checks if character is unconcious but not approached
ln 727 728 update for chage in color variable if chracter is not approached.
ln 749 new lines at the end of echoing the name now. Allows us to apply statuses before or after name.
ln 759 - 780 Updated because cannot is the the text that appears in game not can't