Update TEC Client Scripts.xml
scripts
ln 4427 & 4428 variable creation
ln 4440 enable the update aliase when an update is available.
ln 4450 - 4468 local function collect source file URLS from correct github release. Put them into trackedURLs table.
4488 - 4495: collect URLS when releaseSearch function has found the correct release. Includes error handling.
4510 - 4556 global function trackDownloads. Incompleted fully UNtested.
Parthia initiates it by sending a table with string in it. They contain URLs of files needing download. trackDownloads creates a new table from that. Duplicating the URLs and making a corisponding boolean in table trackedDownloads to mark true when that URL has completed downloading.
After that when a download has completed Parthia sends a string to it, containing a URL of a file that has completed downloading. TrackDownloads that URL as downloaded in table trackedDownloads. it check to see if all URLS have been downloaded. If they have, it will return true. That can then be utilized to run required function after the set of files have been downloaded.
4607 - 4613: collectURLs when player is a beta tester. Includes error collection
4618 - 4624 releaseSearch was changed now includes error throwing. This is additional code for error catching.
4628: nil the table tecGitUpdate.latestJSONTable because it may be large.
4632 - 4638 started the shell for function tecGitUpdate.downloadScripts.