UIMan bugfix
Resolved issue with communication window not resizing when sides are disabled.
Fix issue where if a side was disabled, than the windows that are set to go to that side in the UI layout theme were enabled. They had no where to go. So would go to the top left. They will now look to see if the left or right sides are enabled. If one is they will put the windows in the top of that side. As a by product this makes it so players can disable status and communication windows. Than enable them to go to a tabbed window.
Created easier to understand help command layout. Run `display` or `display help` to see updated display help.
Also found that windows would not save their location. They would go to where the theme told them. Resolved this. Now when a player runs `display save` the location of each window will be saved.
Resolved issue where console windows would show text behind windows above them after windows are enabled or disabled.
Renamed some functions.