bugfix, thoughts filter and side check
scripts
checking for what window side is disabled changed. Top and bottom needed a different detection method than left and right sides.

triggers
`prefs` command has an option to disable colored thoughts. When that option was disabled the trigger to send thoughts to the communications windows would not occur. Created a pattern to detect in that event.