Parthia Map Steps Support
Support for steps or movement that is both vertical and horizontal in a single movement.
This branch has been premered into out current master to avoid conflicts.

Aliases.xml
Added two commands `parthia map steps go up` and `parthia map steps go down`.
Controls if movement that is both vertical and horizontal is up or down.
scripts.xml
9227-9229 just notes
9261 & 9262 create parthiaMap.StepsGoUp and parthiaMap.playerClimbed. With function parthiaMap.config
9394-9436
If the exit the player used does not have a room that appear on the TEC Map. Estimate where the room should be. Display a message if mapping to a mapper can verify the room is in the correct location.
9437 - 9463 If the mapper climbed for this movement is also vertical.  Use parthiaMap.getStepsGoUp to see if mapper needs this to be up or down movement. Display a message to screen that this movement was also vertical. That it was either up or down, and ask the mapper to verify that the correct vertical adjustment was used. Provides a clickable command to change if steps go up or down.
10177-10186 Changed function setEnabled to display debug message of mode even if no message is displayed to screen.
10191-10210 function getStepsGoUp to get the state of variable parthiaMap.stepsGoUp which detects if steps should take the mapper up or down.
10211 - 10249 created function parthiaMap.setStepsGoUp used to set if steps should take mapper up or down.

triggers.xml
Moved and renamed trigger `player climbs` It is not `player climbed` The code within has been remed now it does one thing sets parthiaMap.playerClimbed to true.