mudlet mapper, Simply unique cache table creation
simplified the method to create a cache table for checking if current room is in the unique list.
Only need a list of unique room ids and their names.
Name is a short string requiring few processing cycles to test. No need to store additional data in the cache table. Mudlet mapper database is stored in memory.