mudlet mapper, unique rooms, start changing room method
Got better idea for detecting rooms with ships.
Previous was going to weed out data for a direction. Turned out to be cumbersome. Than would simply not work for the exits that echo on the screen. The word and would be in the wrong place...
Decided to better off making a function that can be manually called. It will store a separate set of user data entries if it is ran and the room looks different that it had previously. Skeleton of function there with notes. Not created yet.
This commit is removing previous changes so the idea is possible.

10749 - 10762 script group `parthiaMap` updated notes
10765 - 10822 function parthiaMap.strMatchDataStore, created function to match whole strings against a map user data store. Created because managing more than one function that does the same thing is a terrible idea.
10857 - 10906, function parthiaMap.checkName, added notes. Better error checking.
Removed lines, removed function parthiaMap.checkCompass and parthiaMap.checkExitsEcho
11053 - 11077, created skeleton functions parthiaMap.changingRoom and parthiaMap.playerLost