mudlet mapper, send full direction name
To resolve issue where TEC can see the movement command "sw". Than if there is no exit to the south west. It will check for special exit names. If one contains a word that starts with sw "swamp" it will move that direction but echo it is walking southwest.

aliases.xml
All movement aliases were updated to use a new function for sending movement commands `sendMovement`.
All patterns are now case insensitive.
Added patterns for limp and crawl movement commands.

keybindings.xml
updated so they send a go fulldirection name. After sending it appears this was not needed. Looks like it checks for aliases with keybindings before sending.
3668, function addLinksToDirections updated it's send command. Also found and resolved a bug in the function. The send was lacking quotes so it was not working.
3998 - 4042, created function sendMovement. It converts movements it receives to full direction names and sends them to the server. If a special exit is used it just sends what is received.
12427-12428, expanded table parthiaMap.directionsTable to include up and down.
12956, updated compass click to use sendMovement function

triggers.xml
found a new movement fail echo "The water is too deep to continue any further in that direction" added it to the failed movement trigger. Moved the regex patterns to the bottom.