Automation now resumes after interrupt completes
When an interrupt completed, it would always have parthiaAutomationCall return true. This is correct behavior for an automation. However interrupts should resume an automation after an interrupt runs.

parthiaIterruptStop now returns true if a previous automation was running that should be resumed.

parthiaAutomationCall when iterating an interrupt. If the interrupt stopped an automation to run. That automation will be resumed.